fix(code): initialize snippets with hidden sidebar
